feat(THEEDGE-3581): clean up; adjust
