+++ src/components/InventoryGroupDetail/InventoryGroupDetail.js
@@ -1,19 +1,28 @@
-import { Bullseye, PageSection, Spinner } from '@patternfly/react-core';
+import {
+  Bullseye,
+  PageSection,
+  Spinner,
+  Tab,
+  Tabs,
+} from '@patternfly/react-core';
 import useChrome from '@redhat-cloud-services/frontend-components/useChrome';
 import PropTypes from 'prop-types';
-import React, { Suspense, lazy, useEffect } from 'react';
+import React, { Suspense, lazy, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchGroupDetail } from '../../store/inventory-actions';
+import GroupSystems from '../GroupSystems';
 import GroupDetailHeader from './GroupDetailHeader';
 import { usePermissionsWithContext } from '@redhat-cloud-services/frontend-components-utilities/RBACHook';
-import { REQUIRED_PERMISSIONS_TO_READ_GROUP } from '../../constants';
-import { EmptyStateNoAccessToGroup } from './EmptyStateNoAccess';
-import HybridInventoryTabs from '../../components/InventoryTabs/HybridInventoryTabs';
 import {
-  PageHeader,
-  PageHeaderTitle,
-} from '@redhat-cloud-services/frontend-components/PageHeader';
-import Main from '@redhat-cloud-services/frontend-components/Main';
+  REQUIRED_PERMISSIONS_TO_READ_GROUP,
+  REQUIRED_PERMISSIONS_TO_READ_GROUP_HOSTS,
+} from '../../constants';
+import {
+  EmptyStateNoAccessToGroup,
+  EmptyStateNoAccessToSystems,
+} from './EmptyStateNoAccess';
+import HybridInventoryTabs from '../../components/InventoryTabs/HybridInventoryTabs';
+import ImmutableDevicesView from '../InventoryTabs/ImmutableDevices/EdgeDevicesView';
 const SuspenseWrapper = ({ children }) => (
   <Suspense
     fallback={
@@ -25,7 +34,8 @@ const SuspenseWrapper = ({ children }) => (
     {children}
   </Suspense>
 );
-const GroupTabDetails = lazy(() => import('./GroupTabDetails'));
+const GroupDetailInfo = lazy(() => import('./GroupDetailInfo'));
+
 const InventoryGroupDetail = ({ groupId }) => {
   const dispatch = useDispatch();
   const { data } = useSelector((state) => state.groupDetail);
@@ -34,6 +44,10 @@ const InventoryGroupDetail = ({ groupId }) => {
 
   const { hasAccess: canViewGroup } = usePermissionsWithContext(
     REQUIRED_PERMISSIONS_TO_READ_GROUP(groupId)
+  );
+
+  const { hasAccess: canViewHosts } = usePermissionsWithContext(
+    REQUIRED_PERMISSIONS_TO_READ_GROUP_HOSTS(groupId)
   );
 
   useEffect(() => {
@@ -49,52 +63,77 @@ const InventoryGroupDetail = ({ groupId }) => {
     );
   }, [data]);
 
+  const [activeTabKey, setActiveTabKey] = useState(0);
+
   // TODO: append search parameter to identify the active tab
+
   return (
     <React.Fragment>
-      <PageHeader className="pf-m-light">
-        <PageHeaderTitle title="Systems" />
-        <GroupDetailHeader groupId={groupId} />
-      </PageHeader>
-
-      <Main>
-        {canViewGroup ? (
-          <HybridInventoryTabs
-            ConventionalSystemsTab={
-              <SuspenseWrapper>
-                <GroupTabDetails {...groupId} />
-              </SuspenseWrapper>
-            }
-            ImmutableDevicesTab={
-              <SuspenseWrapper>
-                <GroupTabDetails {...groupId} />
-              </SuspenseWrapper>
-            }
-            groupName={groupName}
-            groupId={groupId}
-            isImmutableTabOpen={false}
-            isEdgeParityEnabled={true}
-            tabPathname={InventoryGroupDetail.groupTabPathName}
-          />
-        ) : (
-          <PageSection>
-            <EmptyStateNoAccessToGroup />
-          </PageSection>
-        )}
-      </Main>
+      <GroupDetailHeader groupId={groupId} />
+      {canViewGroup ? (
+        <PageSection variant="light" type="tabs">
+          <Tabs
+            activeKey={activeTabKey}
+            onSelect={(event, value) => setActiveTabKey(value)}
+            aria-label="Group tabs"
+            role="region"
+            inset={{ default: 'insetMd' }} // add extra space before the first tab (according to mocks)
+          >
+            <Tab eventKey={0} title="Systems" aria-label="Group systems tab">
+              <PageSection>
+                {canViewHosts ? (
+                  <HybridInventoryTabs
+                    ConventionalSystemsTab={
+                      <SuspenseWrapper>
+                        <GroupSystems groupName={groupName} groupId={groupId} />
+                      </SuspenseWrapper>
+                    }
+                    ImmutableDevicesTab={
+                      <SuspenseWrapper>
+                        <ImmutableDevicesView groupUUID={groupId} />
+                      </SuspenseWrapper>
+                    }
+                    groupName={groupName}
+                    groupId={groupId}
+                    isImmutableTabOpen={false}
+                    isEdgeParityEnabled={true}
+                    tabPathname={'groups'}
+                  />
+                ) : (
+                  <EmptyStateNoAccessToSystems />
+                )}
+              </PageSection>
+            </Tab>
+            <Tab eventKey={1} title="Group info" aria-label="Group info tab">
+              {activeTabKey === 1 && ( // helps to lazy load the component
+                <PageSection>
+                  <Suspense
+                    fallback={
+                      <Bullseye>
+                        <Spinner />
+                      </Bullseye>
+                    }
+                  >
+                    <GroupDetailInfo />
+                  </Suspense>
+                </PageSection>
+              )}
+            </Tab>
+          </Tabs>
+        </PageSection>
+      ) : (
+        <PageSection>
+          <EmptyStateNoAccessToGroup />
+        </PageSection>
+      )}
     </React.Fragment>
   );
 };
 
 InventoryGroupDetail.propTypes = {
   groupId: PropTypes.string.isRequired,
-  groupName: PropTypes.string,
-};
-InventoryGroupDetail.defaultProps = {
-  groupTabPathName: 'insights/inventory/groups',
 };
 SuspenseWrapper.propTypes = {
   children: PropTypes.element,
 };
-
 export default InventoryGroupDetail;
